moved events from separate project to the main one
