+++ axon-spodddify/src/main/java/org/skramer/spodddify/invoice/InvoiceController.java
@@ -5,6 +5,7 @@ import java.util.concurrent.CompletableFuture;
 
 import org.axonframework.messaging.responsetypes.ResponseTypes;
 import org.axonframework.queryhandling.QueryGateway;
+import org.skramer.spodddify.invoice.event.GetAllInvoicesForAccountQuery;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RestController;
