+++ axon-spodddify/src/main/java/org/skramer/spodddify/invoice/InvoiceProjection.java
@@ -6,6 +6,8 @@ import java.util.List;
 
 import org.axonframework.eventhandling.EventHandler;
 import org.axonframework.queryhandling.QueryHandler;
+import org.skramer.spodddify.invoice.event.GetAllInvoicesForAccountQuery;
+import org.skramer.spodddify.invoice.event.InvoiceCreatedEvent;
 import org.springframework.stereotype.Service;
 
 import lombok.AllArgsConstructor;
