+++ axon-spodddify/src/main/java/org/skramer/spodddify/invoice/event/GetAllInvoicesForAccountQuery.java
@@ -1,0 +1,8 @@
+package org.skramer.spodddify.invoice.event;
+
+import lombok.Value;
+
+@Value
+public class GetAllInvoicesForAccountQuery {
+    private String billingAccountId;
+}
