+++ axon-spodddify/src/main/java/org/skramer/spodddify/payment/command/SettleAccountBalanceCommand.java
