+++ axon-spodddify/src/main/java/org/skramer/spodddify/payment/domain/PaymentPlan.java
