+++ axon-spodddify/src/main/java/org/skramer/spodddify/payment/event/BillingAccountCreated.java
@@ -5,7 +5,7 @@ import org.skramer.spodddify.payment.domain.PaymentPlan;
 import lombok.Value;
 
 @Value
-public class BillingAccountCreatedEvent {
+public class BillingAccountCreated {
     private String accountId;
     private long balance;
     private PaymentPlan paymentPlan;
