+++ axon-spodddify/src/main/java/org/skramer/spodddify/payment/event/BillingAccountNotFound.java
@@ -3,6 +3,6 @@ package org.skramer.spodddify.payment.event;
 import lombok.Value;
 
 @Value
-public class BillingAccountNotFoundEvent {
+public class BillingAccountNotFound {
     private String accountId;
 }
