+++ axon-spodddify/src/main/java/org/skramer/spodddify/payment/event/FoundsTransferred.java
@@ -3,7 +3,7 @@ package org.skramer.spodddify.payment.event;
 import lombok.Value;
 
 @Value
-public class FoundsTransferredEvent {
+public class FoundsTransferred {
     private final String billingAccountId;
     private final long balanceChange;
 }
