+++ axon-spodddify/src/main/java/org/skramer/spodddify/payment/event/PaymentPlanChanged.java
