+++ axon-spodddify/src/test/java/org/skramer/spodddify/invoice/InvoiceProjectionTest.java
@@ -8,6 +8,8 @@ import java.util.List;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
+import org.skramer.spodddify.invoice.event.GetAllInvoicesForAccountQuery;
+import org.skramer.spodddify.invoice.event.InvoiceCreatedEvent;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.context.ActiveProfiles;
