+++ settings.gradle
@@ -1,3 +1,2 @@
-include 'spodddify-api'
 include 'axon-spodddify'
