--- spodddify-api/build.gradle
@@ -1,15 +1,0 @@
-apply plugin: 'java'
-
-group 'org.skramer'
-version '0.0.1-SNAPSHOT'
-
-sourceCompatibility = 1.8
-
-repositories {
-    mavenCentral()
-}
-
-dependencies {
-    compile group: 'javax.persistence', name: 'javax.persistence-api', version: '2.2'
-    compileOnly group: 'org.projectlombok', name: 'lombok', version: '1.18.6'
-}
