--- spodddify-api/src/main/java/org/skramer/spodddify/invoice/GetAllInvoicesForAccountQuery.java
@@ -1,8 +1,0 @@
-package org.skramer.spodddify.invoice;
-
-import lombok.Value;
-
-@Value
-class GetAllInvoicesForAccountQuery {
-    private String billingAccountId;
-}
